+++ vscode/test/link-handler.test.ts
@@ -58,6 +58,16 @@ describe('LinkHandler', function () {
             );
         });
 
+        it('should use a short commit hash as the "ref" value when the link type is "commit" and short hashes should be used.', async () => {
+            await setupRepository(root.path);
+
+            sinon.stub(Settings.prototype, 'getUseShortHash').returns(true);
+
+            expect(await createUrl({ url: '{{ ref }}' }, { type: 'commit' })).to.equal(
+                (await git(root.path, 'rev-parse', '--short', 'HEAD')).trim()
+            );
+        });
+
         it('should use the branch name as the "ref" value when the link type is "branch".', async () => {
             await setupRepository(root.path);